requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
fastapi==0.115.12
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
pydantic==2.11.3
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.33.1
    # via pydantic
pyyaml==6.0.2
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
starlette==0.46.1
    # via fastapi
typing-extensions==4.13.2
    # via
    #   -r requirements.in
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
urllib3==2.4.0
    # via requests
uvicorn==0.34.0
    # via -r requirements.in
